♻️ refactor: restructure CherryAgentPage into modular architecture
- Extract utilities into separate modules (formatters, parsers, validators, constants, logProcessors)
- Create custom hooks for state management (useAgents, useSessions, useSessionLogs, useCollapsibleMessages, useAgentExecution)
- Split UI into reusable components (Sidebar, ConversationArea, InputArea, message components, modals)
- Organize styled-components into themed style modules (layout, sidebar, conversation, messages, modals, buttons)
- Reduce main component from ~2400 lines to ~290 lines
- Improve code maintainability and testability with proper separation of concerns
- Follow React best practices with custom hooks and component composition
