+++ src/renderer/src/pages/cherry-agent/components/messages/EmptyConversation.tsx
@@ -1,0 +1,18 @@
+import React from 'react'
+
+import {
+  EmptyConversationComponent,
+  EmptyConversationIcon,
+  EmptyConversationSubtitle,
+  EmptyConversationTitle
+} from '../../styles'
+
+export const EmptyConversation: React.FC = () => {
+  return (
+    <EmptyConversationComponent>
+      <EmptyConversationIcon>💬</EmptyConversationIcon>
+      <EmptyConversationTitle>No messages yet</EmptyConversationTitle>
+      <EmptyConversationSubtitle>Start the conversation below!</EmptyConversationSubtitle>
+    </EmptyConversationComponent>
+  )
+}
