+++ src/renderer/src/pages/cherry-agent/styles/layout.ts
@@ -1,0 +1,36 @@
+import styled from 'styled-components'
+
+export const Container = styled.div`
+  display: flex;
+  flex: 1;
+  flex-direction: column;
+  height: 100vh;
+  width: 100%;
+`
+
+export const ContentContainer = styled.div`
+  display: flex;
+  flex: 1;
+  flex-direction: row;
+  overflow: hidden;
+  height: calc(100vh - 60px); /* Account for navbar */
+`
+
+export const MainContent = styled.div`
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  overflow: hidden;
+  min-width: 0; /* Allow flex shrinking */
+  position: relative;
+`
+
+export const SelectionPrompt = styled.div`
+  flex: 1;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  color: var(--color-text-secondary);
+  font-size: 16px;
+  text-align: center;
+`
